function was renamed from kidneydlseg to dlsegkidney
